added tests for user endpoints
